requirements.txt
# Core dependencies - optimized for Render free tier
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# PDF processing - lightweight
PyMuPDF==1.23.8

# LLM and embeddings - memory optimized
groq==0.11.0
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0+cpu --index-url https://download.pytorch.org/whl/cpu
faiss-cpu==1.7.4

# Essential utilities only
requests==2.31.0
httpx==0.25.2
numpy==1.24.3
loguru==0.7.2

# For better error handling
python-dotenv==1.0.0
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
